not remove openethereum when switching client
