+++ packages/dappmanager/src/index.ts
@@ -5,7 +5,7 @@ import initializeDb from "./initializeDb";
 import { createGlobalEnvsEnvFile } from "./modules/globalEnvs";
 import { generateKeyPair } from "./utils/publickeyEncryption";
 import { copyHostScripts } from "./modules/hostScripts";
-import { deprecateOpenEthereum } from "./modules/ethClient";
+import { switchEthClientIfOpenethereum } from "./modules/ethClient";
 import { runLegacyActions } from "./modules/legacy";
 import { migrateUserActionLogs } from "./logUserAction";
 import { postRestartPatch } from "./modules/installer/restartPatch";
@@ -50,8 +50,8 @@ const server = startDappmanager({
 
 // Deprecate openethereum. MUST be executed before EthClientInstaller daemon to avoid have the error:
 // `Error on eth client installer daemon Error: No client data for target: openethereum`
-deprecateOpenEthereum().catch(e =>
-  logs.error("Error deprecating openethereum", e)
+switchEthClientIfOpenethereum().catch(e =>
+  logs.error("Error switch client openethereum", e)
 );
 
 // Start daemons
